+++ NEWS.md
@@ -1,4 +1,9 @@
-# mpea 0.0.0.90001
+# activeDriverPW 0.0.0.9002
+
+## Major changes
+* Rename to activeDriverPW
+
+# mpea 0.0.0.9001
 
 ## Major Changes
 * Change column contribution method. Column contribution is now reported as the